Switch http_archive demo to use tags (#3)
We can use git tags to avoid having to look up commits corresponding to
a given LLVM revision. The tag format is `llvm-project-${LLVM_COMMIT}`.
I originally tried `llvm/llvm-project@${LLVM_COMMIT}` to make something
that is both autolinked by GitHub and unambiguous in the commit it
references but having URL-encoded characters seemed not worth it.