Document demos and move http-archive-demo to its own branch (#8)
This isn't strictly necessary (like it is with the submodule-demo), but
having the same setup here makes sense. Because this has to reference
the commit it's using by SHA it's not actually possible to have it use
the current commit. I want to set up automation to make this always get
updated to match `main` but not pollute the main branch with lots of
automated commits.