Automatically integrate new LLVM commits (#10)
When a change is pushed to `main`, it will be merged into
`auto-integrate` and then all LLVM commits after that point will get a
corresponding llvm-bazel commit tagged with `llvm-project-${COMMIT}`
that contains the corresponding build files.

Subsequent patches will make this trigger on new LLVM commits as well
(or on a cron) and try pulling in MLIR build file patches from TF.

The intended workflow is that we get candidates for all LLVM commits
out quickly. Users can with minimal delay reference a candidate set of
build files by using only the LLVM commit SHA. When we detect that some
LLVM commit requires a manual intervention, we push a new commit to
main with the submodule updated to that point. The automation then
creates new commits with corresponding tags for the following LLVM
commits (potentially overwriting the commit a tag pointed to). This
means that tags for commits on the auto-integrate branch are not
guaranteed to always point to the same commit and are therefore likely
not suitable for use in `http_archive` since the sha256 would be
different. Once a tag is applied to the main branch, we don't change it
however.

Tested:
You can see the results of this on the current `auto-integrate` branch.
I temporarily modified the workflow to also run on pushes to this PR
branch and it created a series of commits on the auto-integrate branch
with tags. I will drop these commits before merging this PR and let
them get generated from `main`.