Begin filling in a hand-written port of LLVM to Bazel.
This replaces the automatically generated BUILD file with rules directly
written against the LLVM project. Rules are designed to use narrow globs
to be reasonably resilient to simple source additions and removals.
Headers and layering are strictly enforced here. There is significant
logic to build the targets, with some libraries being merged due to
`#include` cycles that cannot be split.

Not all of the LLVM libraries are included yet, just enough to build the
`opt` and `llc` tools, including most of the targets, and the testonly
libraries and tools that include tricky parts of GoogleTest and
GoogleMock.

Still lots of cleanup to do, and then expanding this to cover the entire
LLVM library collection.