Switch to a hand-written port of the LLVM build to Bazel.
This replaces the automatically generated BUILD file with rules directly
written against the LLVM project. Rules are designed to use narrow globs
to be reasonably resilient to simple source additions and removals.
Various tricks are used to minimize repetitive rules around LLVM
targets and help ensure this is reasonably scalable.

Configuration is handled by a template substitution technique that will
directly break when new configuration macros are added to `config.h` or
`llvm-config.h`. Unused entries won't result in an immediate error
though and so may need to be GC-ed periodically. The key is that new
configuration knobs can't silently produce unexpected results: they
require a choice to be made. However, this configuration system can and
should be factored out of the main `BUILD` file and into a separate
Starlark rule in a follow-up. The goal is to keep it very simple and
based on Starlark rather than falling back to other systems. Once that
happens, we will also need to add `select`s and other hooks to configure
the values used as currently they largely hard code a reasonably vanilla
Linux environment but are not portable to other environments.

Headers and layering are strictly enforced. There is significant logic
to build the targets, with some libraries merged (compared to upstream)
to support `#include` cycles that cannot be split.

This includes essentially all of the main LLVM libraries and binary
tools, including those used for testing LLVM itself. It should be enough
to support MLIR and Clang.

However, this doesn't yet add unittests or other testing. Those will
show up in subsequent commits.

There are a few other cleanups needed internally around warning flags,
host/target configuration hooks, and version number generation.

Despite these limitations, this is enough to be reasonably functional
for users of these libraries on Linux, and we're hoping to close at
least the most critical of these gaps quite soon.

Lots of thanks to a bunch of folks at Google who have helped contribute
to the baseline of these rules over the years that I used to produce the
Bazel version, and to the code review on this version!