Move warning configuration from the BUILD file to the `.bazelrc`.
Also enhances the `.bazelrc` to allow more easy build and test with both
Clang and GCC.

I've added some documentation to avoid simple mistakes, and removed
special handling of `NDEBUG` (which Bazel reliably sets with
`--compilation_mode=opt`).

The cleanups to the `.bazelrc` are somewhat opinionated on my part. ;]
It seems really good to make it easy to test different build
configurations here and so getting something a bit more vanilla (rather
than forcing `-c opt` by default) seems good. If we want optimizations
enabled by default, a better way to do that is a custom toolchain that
sets up fastbuild with the desired flags, but for this project it
doesn't seem worth it.