Enable vulkan support (#23)
Mostly adopts the configuration used in https://github.com/google/iree.

The specific `WORKSPACE` configuration of this (as in all things) is
just one example of how to do this. The key thing is that in order to
use the VulkanRuntime target, users must create appropriate
`@vulkan_headers` and `@vulkan_sdk` repo rules. I'm not aware of an
easy way to make this configurable (not have it rely on a specific
name).

We leave all the targets that rely on external dependencies as "manual"
so that users do not need to configure any additional dependencies to
use llvm-bazel.

TESTED=`bazel build @llvm-project//...`