Add support for linking against terminfo.
I'm not aware of an effective external Bazel-style dependency, so this
only adds support for detecting and configuring a system-library. This
is somewhat more complex than other libraries because there are a few
different libraries that provide these APIs.

I've tried to ensure sufficient configuration hooks for these rules to
allow different users of LLVM to enable or not this functionality.