Configure RBE (#36)
This allows remotely executing and caching actions.

The primary motivation is to be able to have a cluster of build agents
all sharing a cache.

We have one issue with using the default RBE toolchain: it doesn't have
xml2-dev, so we get link failures for WindowsManifest. For now, I've
marked the target manual. I also played around with the
`no-remote-exec` tag, but there are two issues: It doesn't actually
*work* without an experimental flag
https://github.com/bazelbuild/bazel/issues/7988. Even turning that on,
Bazel tries to look for the clang binary at the path it would be on the
remote on the local machine, which doesn't work so well.

> src/main/tools/linux-sandbox-pid1.cc:426:
"execvp(/usr/local/bin/clang, 0x1f81180)": No such file or directory

To get this working, we can use a custom docker container for our
toolchain that includes the right dependencies. This isn't too hard to
set up. But before I do it, do we actually care about this binary right
now? I can punt if we don't.