Rework llvm configuration (#48)
The idea of string-substituting the .cmake file was to keep things
simple and ensure that we detected changes to the .cmake file that we
would need to pick up.

As we need to move to selects based on platform this is going to get
super verbose. We've also already pulled some things out into Bazel
defines (e.g. zlib and terminfo features).

As for detecting drift, it only sort of worked, since it wouldn't
detect if a variable was removed, e.g.
https://github.com/llvm/llvm-project/commit/31e5f7120bdd. This leaves
us with some confusing tech debt for later.

Instead this change splits the jobs of setting the appropriate defines
between hardcoded `.h` files, for things that don't change based on
platform, and Bazel defines based on selects, for those that do. We
could eventually move everything into Bazel defines if we wanted to,
but this is maybe more work for things that don't change.

To kick things off, I extracted a few variables that I know will be
different on Windows, but I have no confidence that this is now
configured such that a Windows build would pass and I don't have a box
to test on.

Another approach would be to write a true equivalent to CMake's
`configure_file` and use that along with a dict for the defines.
@chandlerc expressed concerns about their being a python dependency in
the actual build stage, so this should probably be written in C[++].

For detecting drift we move from something that sort of does so to
something that definitely will: literal change-detector tests. To avoid
a loading time dependency on `bazel_skylib` in the actual project's
build rules (that we expect dependent projects to use), these tests are
in the root workspace.