Configuration for a windows build
This now builds with warnings-as-errors (which requires disabling quite
a few errors).

I am by no means a Windows expert. I just got my hands on a Windows VM
and fixed things until everything built. Requires
https://reviews.llvm.org/D89760 and https://reviews.llvm.org/D89758
to build cleanly.

There are a lot of places where we set linkopts iff not windows. Might
want to factor that into a macro, but I decided to wait till we had more
configurations to see if that would remain the differentiator.