Simplify the use of `CC` and fix a bug in it.
The terminfo detection needs to find a working C compiler, but the
strategy used was a bit more complex than needed and didn't work when
the environment variable was set to an absolute path. Fix that by simply
using that path if specified.

This removes any attempt to fallback in the face of a broken `CC`
environment variable -- that is unlikely to be helpful.