Clean-up and remove hard-coded flag from `tblgen` Bazel rule. (#71)
The hard coded flag disables the use of long string literals. This isn't
supported for all tablegen tools, and isn't desirable in any of them. It
is a workaround for compilers that can't handle the long string
literals. For compilers that can handle these literals, using them is
a (significant) compile time improvement.

I don't believe any compilers we're supporting with Bazel currently
require this flag, and if they do we should find out and pass it very
narrowly for those compilers. First step is to remove it here.

I've tested this both on a recent Clang and GCC 5 successfully.