Add Clang and its unittests to the build. (#72)
This covers both the libraries and main binaries for Clang, including
many of the test binaries and a few tools. Much like the LLVM `BUILD`,
it should be sufficient for users that are primarily interested in
linking and using the Clang libraries themselves and isn't sufficient to
build a complete toolchain.

I've also added the unittests which turned out to be very simple
(similar to the LLVM unittests). Again, this should very closely match
the experience of library users of Clang.

I've also built the `clang` executable and successfully compiled the
world's smallest test program or two, so it seems to be in working
order.

So far, only tested on Linux, but with both a Clang and GCC build.

Co-authored-by: Geoffrey Martin-Noble <gcmn@google.com>