Remove the `includes` directive from `tblgen` libraries. (#82)
This caused every single directory tree with a tblgen output in it to be
in the `-isystem` header search path. That is far from desirable, and
was hiding a large number of issues reaching both generated files and
source files with `#include`.

On MacOS it also manifested in an amazing problem where a header in
Clang's AST library called `Availability.h` ended up being used instead
of an internal header by the same name inside of the MacOS libc to
provide system availability macros to the libc headers. The resulting
errors were... fun.

To get these generated files to be found, this change adds
`strip_include_prefix` where necessary to make the relevant headers
available at the expected paths. This is much more narrow and focused.
Making it fully work requires two other workarounds. First, we need to
introduce extra globs or libraries to provide the same headers through
different include paths in a few places that were being hidden by the
`includes` from `tblgen`. Second, we have to switch to `hdrs` instead of
`textual_hdrs` in numerous places to work around a Bazel bug:
https://github.com/bazelbuild/bazel/issues/12424

However, with this, we get the Bazel side to a clean, principled
implementation of the requirements of LLVM.