Remove most of the remaining include search hacks. (#83)
Notably, all of the `-Iexternal/...` tricks are gone through the careful
application of extra entries in `srcs` to enable
file-relative-inclusion, and special header-only libraries which
re-expose header files with the desired prefixes using combinations of
`strip_include_prefix` and in one case `include_prefix`.

The result likely has some missing dependencies that will only be
detected if we enable header parsing in the future. I didn't want to try
to figure those out from scratch in every case without the tooling to
tell me if I'm wrong.

There are also still places where we should be doing intentional textual
headers, but cannot due to the Bazel bug that prevents
`strip_include_prefix` from working with them. I've left appropriate
comments about fixing those.

Even with the workarounds required here, I think this is a positive
step, and I don't really want to block on cleaning up all the code that
this copes with.

I also had to fix some of the build issues with the `llvm-exegesis`
stuff that I hadn't noticed before in order to test all of this, but
I did and that's included. The unittests and everything for that builds
now if you happen to have the correct stuff installed locally.