Add missing BUILD rules for the OpenACC frontend library.
These are only used by one unittest and the Flang frontend currently.
Somehow, the unittest in question compiles and runs successfully on all
the Linux hosts I've tested -- I really don't know how. My best guess is
that it finds the necessary generated header from an LLVM installation
that is made visible even through the sandbox to the build.

However, on MacOS, this doesn't work and the problem surfaces. Fix this
the right way by adding the missing build infrastructure matching the
OMP infrastructure (but simpler given the code simplicity of this
library).

With this, all of the LLVM and Clang unittests build and pass for me on
my Mac laptop.