Move linkopts to llvm:Support (#189)
`-lm` is needed by Support itself. Adding this on every possible leaf
binary seems a bit silly. There will be some weirdness by nature of
this being a transitive thing (line `includes`) that means we might
have some things happen to work because they dep on Support even though
there are other libraries that really should have those options as well
and don't have them, but I'm not sure how to avoid that. Luckily the
linker error gives you the name of the object file, so I think adding
the option on the appropriate library if/when that comes up shouldn't
be too bad.

Fixes #188