Allow configuring the set of LLVM targets built.
For library users that only need specific targets built, this is
particularly significant both to speed up builds and also avoid
embedding large amounts of unnecessary code.

This somewhat surfaced that the repository rule being the overlay
creation doesn't really make sense, and so I've folded that into
a single repository rule where it is reasonable to add LLVM-specific
logic to fully customize things. Happy to use a different structure if
folks have preferences.

I also know that all of this is on the way upstream, but I'd like to use
this ASAP, and didn't want to necessarily gate on the upstream patches
landing if that's OK.